src/security_system/inference_engine/weaponCategory/EfficientNet/requirements.txt
torch==1.10.0
torchvision==0.11.1
numpy==1.21.2
pandas==1.3.3
Pillow==8.4.0
matplotlib==3.4.3
scikit-learn==0.24.2
argparse==1.4.0
tensorboard==2.7.0
